Make it clear that all protocol ops act on bytestrings
Running the tests as-was in Python 3.x results in a number of errors
because in Python 2.x the literal

    'foo'

is a bytestring, whereas it is a character sequence (a unicode object)
in Python 3.x. All protocol-level operations act on bytestrings, and
indeed raise errors if they are passed non-bytestrings, so this commit
makes liberal use of the literal form

    b'foo'

to make this clear. This is compatible with Python 2.6 and above.